requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
docopt==0.6.2
idna==3.3
lxml==4.8.0
mypy==0.950
mypy-extensions==0.4.3
pymongo==4.1.1
python-dotenv==0.20.0
requests==2.27.1
soupsieve==2.3.2.post1
tomli==2.0.1
types-requests==2.27.25
types-urllib3==1.26.14
typing_extensions==4.2.0
urllib3==1.26.9
yarg==0.1.9